Enable browser expiring and default sessions ("keep me signed in")
This also fixes an issue where roundcube sends it's session cookie to
MiaB and vise-versa, because both had a cookie path of "/". This is
potentially a security issue if other web applications are installed
on the host. The bug is present in the current version MiaB.

Set a browser caching policy for max-age for static files.

Remove the requirement a user be logged in to access the logout endpoint.